Doc changes (#32)
* Update README

* Update README

* Parquet files are typed so no need to cast timestamp

* More changes

* Update README.md


* Doc updates

* Update documentation, set to 'trigger', optional None in constraints, reorg loading in __main__

* Docs

* Add test for invalid constraints + coderabbit suggestions

* Configs

* Remove e2e test from meds

* Predicate count type

* Using int64

* Missed some uint16s

* Check for empty df after getting trigger event

* Drop null timestamps and subject_ids

* Fix tests

* MEDS should be aggregated by subject id and timestamp

* Add index_timestamp column

* Fix tests and configs

* Fix doctests

* Tests for index_timestamp and label

* Test for duplicate timestamps

* Selecting columns for index_timestamp and label

* Unique check

* aggregate in csv and check for unique in query

* Coderabbit fixes

* Refactor + initial e2e test

* rename

* Other changes

* Making the hydra CLI work (#33)

* This made the CLI work for me

* This works

* test doesn't work but removing unused files

* Config and script modifications

* Adjusted direct input format and config.

* Moved error handling to predicates for file existence

* Fix doctest

* integration test finally working

---------


* Update python version in tests

---------

+++ src/aces/types.py
@@ -10,7 +10,7 @@ from datetime import timedelta
 import polars as pl
 
 # The type used for final aggregate counts of predicates.
-PRED_CNT_TYPE = pl.UInt16
+PRED_CNT_TYPE = pl.Int64
 
 # The key used in the endpoint expression to indicate the window should be aggregated to the record start.
 START_OF_RECORD_KEY = "_RECORD_START"
